Move modulus.utils.sfno.zenith_angle to modulus.utils.zenith_angle
This gives better visibility
